refactor: 패키지 의존 방향 통일을 위한 레포지토리 함수 이동
