+++ backend/src/test/java/reviewme/question/repository/OptionItemRepositoryTest.java
@@ -1,7 +1,6 @@
 package reviewme.question.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
 import static reviewme.fixture.OptionGroupFixture.선택지_그룹;
 import static reviewme.fixture.OptionItemFixture.선택지;
 import static reviewme.fixture.QuestionFixture.선택형_필수_질문;
@@ -64,28 +63,5 @@ class OptionItemRepositoryTest {
 
         // then
         assertThat(actual).containsExactlyInAnyOrder(optionItem1, optionItem2, optionItem3);
-    }
-
-    @Test
-    void 질문_아이디에_해당하는_모든_옵션_아이템을_불러온다() {
-        // given
-        Question question1 = questionRepository.save(선택형_필수_질문());
-        Question question2 = questionRepository.save(선택형_필수_질문());
-        OptionGroup optionGroup1 = optionGroupRepository.save(선택지_그룹(question1.getId()));
-        OptionGroup optionGroup2 = optionGroupRepository.save(선택지_그룹(question2.getId()));
-
-        OptionItem optionItem1 = optionItemRepository.save(선택지(optionGroup1.getId()));
-        OptionItem optionItem2 = optionItemRepository.save(선택지(optionGroup1.getId()));
-        OptionItem optionItem3 = optionItemRepository.save(선택지(optionGroup2.getId()));
-
-        // when
-        List<OptionItem> optionItemsForQuestion1 = optionItemRepository.findAllByQuestionId(question1.getId());
-        List<OptionItem> optionItemsForQuestion2 = optionItemRepository.findAllByQuestionId(question2.getId());
-
-        // then
-        assertAll(
-                () -> assertThat(optionItemsForQuestion1).containsOnly(optionItem1, optionItem2),
-                () -> assertThat(optionItemsForQuestion2).containsOnly(optionItem3)
-        );
     }
 }
